gradio-ui/requirements.txt
gradio==4.15.0
requests==2.31.0
numpy==1.24.4
Pillow==10.2.0
